+++ src/components/Admin/index.ts
@@ -1,10 +1,7 @@
 export * from './AdminNav';
 export * from './AdminPage';
-export * from './GroupEdit';
-export * from './GroupList';
 export * from './GroupPage';
-export * from './User';
-export * from './UserEdit';
+export * from './ListPage';
 export * from './UserForm';
 export * from './UserList';
 export * from './UserPage';
