--- src/utils/mapChallenges.ts
@@ -1,4 +1,0 @@
-export const mapChallenges = (challenge: { textID: string; name: string }) => ({
-  id: challenge.textID,
-  value: challenge.name,
-});
