--- src/utils/mapGroups.ts
@@ -1,4 +1,0 @@
-export const mapGroups = (group: { id: string; name: string }) => ({
-  id: group.id,
-  value: group.name,
-});
